Moved computerPlay() into the playRound() function and added gameplay buttons
Not sure if nesting computerPlay() into playRound() is best practice but
it makes sense.

I learned that .addEventListener() requires 'event' and 'function' as
parameters. This means the second parameter should preferably not contain
parameters as doing so would invoke a function call and return a value which
is passed into .addEventListener(), instead of passing the function itself
into the call.

Therefore to work around this I used an anonymous function containing
playRound(playerSelection) for each of the different button.addEventListener
calls. This means the add.EventListener() does indeed receive a function
object as a parameter, which will not have to be invoked immediately. The
game now works as intended - playRound() is only invoked when the
"click" event occurs.